Application/requirements.txt
django 
djangorestframework
djangorestframework_simplejwt
drf_yasg
